Added custom conversion handling for csv importing.
