+++ Csv/CsvCharacterMap.cs
@@ -1,11 +1,16 @@
-﻿namespace w3_assignment_ksteph.Csv;
+﻿
+using CsvHelper;
+using w3_assignment_ksteph.Character;
+using w3_assignment_ksteph.Csv;
+using CsvHelper.Configuration;
 
+namespace w3_assignment_ksteph.Csv;
 // The Character Map is used to convert character information from csv format to a Character object.
-public class CsvCharacterMap : CsvClassMap<Character>
+public class CsvCharacterMap : ClassMap<Character.Character>
 {
     public CsvCharacterMap()
     {
-        InventoryConverter inventoryConveter = new InventoryConverter()
+        InventoryConverter inventoryConverter = new InventoryConverter();
     
         Map(character => character.Name);
         Map(character => character.Class);
