+++ Inventory/Inventory.cs
@@ -1,4 +1,6 @@
-﻿namespace w3_assignment_ksteph.Inventory;
+﻿using w3_assignment_ksteph.Csv;
+
+namespace w3_assignment_ksteph.Inventory;
 public class Inventory
 {
     public List<Item>? Items { get; set; } = new();
@@ -8,4 +10,6 @@ public class Inventory
     {
         Items = items;
     }
+
+    public override string ToString() => InventorySerializer.Serialize(this)!;
 }
