+++ Inventory/InventoryManager.cs
@@ -6,5 +6,5 @@ using w3_assignment_ksteph.Csv;
 
 public class InventoryManager
 {
-    public static void ListInventory(Inventory inventory) => InventorySerializer.ListInventory(ToString(inventory));
+    public static void ListInventory(Inventory inventory) => InventorySerializer.ListInventory(inventory);
 }
